+++ src/textual/_node_list.py
@@ -119,6 +119,26 @@ class NodeList(Sequence["Widget"]):
                 self._nodes_by_id[widget_id] = widget
             self._updates += 1
 
+    def _move_to_end(self, widget: Widget) -> None:
+        """Move a widget to the end.
+
+        Args:
+            widget: Widget to move (should be in the list).
+        """
+        if widget in self._nodes_set:
+            self._nodes.remove(widget)
+            self._nodes.append(widget)
+            self._updates += 1
+
+    def _replace(self, widget: Widget, new_widget: Widget) -> bool:
+        try:
+            index = self._nodes.index(widget)
+        except ValueError:
+            return False
+        else:
+            self._nodes[index] = new_widget
+        return True
+
     def _ensure_unique_id(self, widget_id: str) -> None:
         if widget_id in self._nodes_by_id:
             raise DuplicateIds(