+++ src/textual/dom.py
@@ -310,6 +310,21 @@ class DOMNode(MessagePump):
             self.call_later(self._initialize_data_bind)
         return self
 
+    def copy_state(self, node: DOMNode) -> None:
+        """Copy the state from another node.
+
+        The default implementation will copy over reactives with the same name,
+        but this method may be override for different strategies.
+
+        Args:
+            node: A node to copy state from.
+        """
+        for key, reactive in node._reactives.items():
+            if reactive.state:
+                if key in self._reactives:
+                    if hasattr(node, reactive.internal_name):
+                        setattr(self, key, getattr(node, key))
+
     def _initialize_data_bind(self) -> None:
         """initialize a data binding.